+++ lib/src/view/_.dart
@@ -4,7 +4,7 @@ export 'dialog/_.dart';
 export 'model/_.dart';
 export 'page/_.dart';
 export 'painter/scrawl_painter.dart';
-export 'res/gaps.dart';
+export 'res/_.dart';
 export 'widget/_.dart';
 export 'toast.dart';
 export 'typedef.dart';